Move `compare_mesh_with_image` to `.utils`
