+++ nanomesh/__init__.py
@@ -4,7 +4,7 @@ import sys
 from ._tetgen_wrapper import tetrahedralize
 from ._triangle_wrapper import triangulate
 from .image import Plane, Volume
-from .image2mesh import Mesher2D, Mesher3D, compare_mesh_with_image
+from .image2mesh import Mesher2D, Mesher3D, plane2mesh, volume2mesh
 from .mesh import LineMesh, TetraMesh, TriangleMesh
 from .mesh_container import MeshContainer
 from .region_markers import RegionMarker
@@ -21,16 +21,17 @@ __all__ = [
     '__author__',
     '__email__',
     '__version__',
-    'compare_mesh_with_image',
     'LineMesh',
+    'MeshContainer',
     'Mesher2D',
     'Mesher3D',
-    'MeshContainer',
     'Plane',
-    'triangulate',
+    'plane2mesh',
+    'RegionMarker',
     'tetrahedralize',
     'TetraMesh',
-    'RegionMarker',
     'TriangleMesh',
+    'triangulate',
     'Volume',
+    'volume2mesh',
 ]
