+++ nanomesh/image2mesh/__init__.py
@@ -1,4 +1,4 @@
-from .mesher2d import Mesher2D, compare_mesh_with_image, plane2mesh
+from .mesher2d import Mesher2D, plane2mesh
 from .mesher3d import Mesher3D, volume2mesh
 
 __all__ = [
