+++ nanomesh/image2mesh/mesher2d/__init__.py
@@ -1,9 +1,8 @@
-from ._helpers import compare_mesh_with_image, pad
+from ._helpers import pad
 from ._mesher import Mesher2D, plane2mesh
 from ._polygon import Polygon
 
 __all__ = [
-    'compare_mesh_with_image',
     'plane2mesh',
     'Mesher2D',
     'pad',
