+++ nanomesh/image2mesh/mesher2d/_helpers.py
@@ -2,47 +2,13 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, List
 
-import matplotlib.pyplot as plt
 import numpy as np
 from scipy.spatial.distance import cdist
 
 from nanomesh.region_markers import RegionMarker
 
 if TYPE_CHECKING:
-    from nanomesh.mesh import LineMesh, TriangleMesh
-
-
-def compare_mesh_with_image(image: np.ndarray,
-                            mesh: TriangleMesh,
-                            cmap: str = None,
-                            **kwargs):
-    """Compare mesh with image.
-
-    Parameters
-    ----------
-    image : 2D array
-        Image to compare mesh with
-    mesh : TriangleMesh
-        Triangle mesh to plot on image
-    cmap : str
-        Matplotlib color map for `ax.imshow`
-    **kwargs :
-        Extra keyword arguments passed on to plotting function.
-
-    Returns
-    -------
-    ax : matplotlib.Axes
-    """
-    fig, ax = plt.subplots()
-
-    mesh.plot_mpl(ax=ax, **kwargs)
-
-    ax.imshow(image, cmap=cmap)
-    ax.axis('image')
-    ax.set_xticks([])
-    ax.set_yticks([])
-
-    return ax
+    from nanomesh.mesh import LineMesh
 
 
 def pad(
